@@ -14,14 +14,14 @@ classifiers = [
   "Operating System :: OS Independent",
   "Programming Language :: Python",
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
   "Programming Language :: Python :: 3.13",
+  "Programming Language :: Python :: 3.14",
 ]
 readme = "README.rst"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
   "ckanapi",
   "click",
@@ -75,7 +75,7 @@ exclude_also = [
 version.source = "uv-dynamic-versioning"
 
 [tool.mypy]
-exclude = ["build/"]
+files = ["doc", "transport_data"]
 
 [[tool.mypy.overrides]]
 module = [
