@@ -2,7 +2,7 @@
 
 import logging
 import re
-from typing import TYPE_CHECKING, List, Optional, Tuple, cast
+from typing import TYPE_CHECKING, List, Tuple, cast
 
 from sdmx.model import common, v21
 
@@ -236,7 +236,7 @@ def read_worksheet(
     ws: "Worksheet",
     msd: "v21.MetadataStructureDefinition",
     cs_dims: "v21.ConceptScheme",
-) -> Optional["v21.MetadataReport"]:
+) -> "v21.MetadataReport | None":
     """Read a metadata report from the worksheet `ws`.
 
     Parameters
@@ -248,7 +248,7 @@ def read_worksheet(
     from transport_data.org.metadata import _get
 
     # Mapping from names (not IDs) to MetadataAttributes
-    mda_for_name = {
+    mda_for_name: dict[str, "common.MetadataAttribute"] = {
         str(mda.concept_identity.name): mda
         for mda in msd.report_structure["ALL"].components
     }
@@ -275,7 +275,8 @@ def read_worksheet(
     mdr = v21.MetadataReport()
     mdr.attaches_to = tok
 
-    mda = None  # Reference to the MetaDataAttribute describing the current row
+    # Reference to the MetaDataAttribute describing the current row
+    mda = common.MetadataAttribute()
     dimension_concepts = []
 
     # Iterate over rows in the worksheet, skipping the first
