@@ -4,6 +4,7 @@ from typing import TYPE_CHECKING
 
 import pytest
 import sdmx
+from sdmx.message import DataMessage
 from sdmx.model import v21
 
 from transport_data import Config, testing
@@ -44,16 +45,22 @@ def test_convert(caplog, rtdb_test_data: "pathlib.Path") -> None:
         return obs.dimension["GEO"].value == value
 
     for geo, N in COUNTRIES:
+        # Create a DataMessage to work around https://github.com/khaeru/sdmx/issues/251
+        dfd = v21.DataflowDefinition(structure=ds1.structured_by)  # type: ignore [arg-type]
+
         ds2 = v21.DataSet(
+            described_by=dfd,
             structured_by=ds1.structured_by,
             obs=list(filter(partial(_filter_geo, geo), ds1.obs)),
         )
+        dm = DataMessage(dataflow=dfd, data=[ds2])
 
-        df = sdmx.to_pandas(ds2).reset_index()
-        if len(df):
-            df = df.query("value > 0")
+        if not len(ds2):
+            continue
 
-        assert N == len(df)
+        df = sdmx.to_pandas(dm).reset_index()
+
+        assert N == len(df.query("value > 0 "))
 
 
 @pytest.mark.xfail(
